Add Features and Update UI
Add features:
- Add movie/series to favorite
- Add movie/series to watchlist
- Update UI (show detailed scoring each tv/movies on detail page (*API limitation), retouch on favorite page and watchlist page, add progress bar)

+++ app/src/main/java/com/waffiq/bazz_movies/data/remote/response/MovieTvResponse.kt
@@ -50,7 +50,7 @@ data class ResultItem(
 	val popularity: Double? = null,
 
 	@SerializedName("vote_average")
-	val voteAverage: Double? = null,
+	val voteAverage: Float? = null,
 
 	@SerializedName("name")
 	val name: String? = null,
@@ -76,3 +76,4 @@ data class ResultItem(
 	@SerializedName("release_date")
 	val releaseDate: String? = null
 ): Parcelable
+
